Guessing game: For two human players
Create a multiplayer game!!

Player 2 looks to the other direction, while player 1 enters a secret number between 1 and 100. (use input() to get a number from the user)

The screen is cleared. (Print a lot of blank lines with "\n" * 100)

Player 2 returns and tries to guess the number. The computer might say Correct! You win! and the game is done.

Otherwise print Your guess is too high! or Your guess is too low! and ask again.