Simple Square
The following program prints the word hello 5 times and the word goodbye 8 times:

n = 5
for i in range(n):
    print("hello")

n = 8
for i in range(n):
    print("goodbye")
Exercise: Create a program that prints a square of n*n asterisks.

For example, if the first line in the program is

n =  3
it will produce the following output:

***
***
***
Once your program works, modify the first line and make sure it prints a square of the correct size. For example, changing the first line to:

n =  8
should print a 8*8 square.