Sets vs Lists
Assuming we have a big list of random numbers:

>>> import random 
>>> mylist = random.sample(range(1, 1000001), 10000)
>>> type(mylist)
list
>>> len(mylist) 
10000
And the same data in a set:

>>> myset = set(mylist)
>>> type(myset)
set
>>> len(myset) 
10000
What will be faster, looking up a value in the set or in the list?

x = random.randint(1, 1000000)
x in mylist
x in myset
Let's check it out!

Create a program that:

Creates a set and a list of 10,000 random numbers between 1 and 1,000,000 (code above).
Measure how much time it takes to repeats the lookup above 10,000 times, with the list.
Measure how much time it takes to repeats the lookup above 10,000 times, with the set.
Use time.time() to measure how much time this takes:

import time
print(time.time())
Which method is faster? Why?