Calculate Median (Part 2)
Create a program that accepts numbers from the user, until a blank line is entered:

> 10
> 15
> 20
> -40
> 90
>
And then prints the number of items, minimum, maximum, average and median of the numbers entered. (The median is the value of the middle number after sorting them. In case of even items, return the average of the two middle items):

Items: 5
Min: -40
Max: 90
Avg: 19.0
Median: 15